bot/requirements.txt
aiogram==3.20.0
aiosqlite==0.21.0
aiohttp==3.11.18
python-dotenv==1.1.0
mcrcon~=0.7.0
Jinja2~=3.1.6
docker==7.1.0
apscheduler==3.11.0
